split off TestMultiIndexPanel class
